Day 9
Squashed commit of the following:

commit 0fc1ccf46e1c638a1d5f39a60c3d5b23300a127a
Author: Joaquín Monteagudo Gómez <Kindos7@gmail.com>
Date:   Fri Oct 17 00:37:36 2025 +0200

    Format code

commit 4868e519baa35392060853c53bbd95b4f0a12c59
Author: Joaquín Monteagudo Gómez <Kindos7@gmail.com>
Date:   Fri Oct 17 00:34:23 2025 +0200

    Day 9 part 1 and 2, working

commit 7ba1ea9944092d4015b79205be43603abb4a2a0f
Author: Joaquín Monteagudo Gómez <Kindos7@gmail.com>
Date:   Fri Oct 17 00:21:47 2025 +0200

    Expand memory and IntCell sizes

commit a931b69f3fdb1bc586c0d85151a8cbc4318e29f0
Author: Joaquín Monteagudo Gómez <Kindos7@gmail.com>
Date:   Fri Oct 17 00:14:43 2025 +0200

    Implement relative base offset

commit 0c499652403162bba00a5be8f4e704f8bdcc496e
Author: Joaquín Monteagudo Gómez <Kindos7@gmail.com>
Date:   Thu Oct 16 23:59:29 2025 +0200

    Add relative parameter support

+++ src/types.rs
@@ -1,1 +1,1 @@
-pub type IntCell = i32;
+pub type IntCell = i64;
